+++ app/Repositories/Eloquent/Admin/Buses/BusesRouteRepository.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Repositories\Eloquent\Admin\Buses;
 
+use App\Models\UsersModel\Buses\Buses;
 use App\Models\UsersModel\Buses\BusesRoute;
 use App\Repositories\Eloquent\Repository;
 
@@ -80,11 +81,21 @@ class BusesRouteRepository extends Repository {
     }
     /*删除班车线路*/
     public function destroyBusesRoute($id){
-        $result = $this->delete($id);
-        if ($result) {
-            flash('删除成功','success');
-        } else {
-            flash('删除失败','error');
+        $result = false;
+        if($this->model->where('buses_pid',$id)->exists()){
+            flash('有子线路未删除,请先删除所有子线路！','success');
+        }else{
+
+            if(Buses::where('busesroute_id',$id)->exists()){
+                flash('此线路下有班车,须清理班车后删除此线路！','success');
+            }else{
+                $result = $this->delete($id);
+                if ($result) {
+                    flash('删除成功','success');
+                } else {
+                    flash('删除失败','error');
+                }
+            }
         }
         return $result;
     }
